+++ audio/common/audio_entry.h
@@ -9,6 +9,6 @@
 
 void *audio_control_init(uint8_t thread_count);
 void audio_control_loop(void *context);
-void audio_process_data(void *context, int thread_id);
+void audio_process_data(void *context, uint8_t thread_id);
 
 #endif /* _AUDIO_ENTRY_H_ */
